rag_apps/vision_rag/requirements.txt
streamlit
cohere
requests
PyMuPDF
Pillow
numpy
scikit-learn
python-dotenv 